+++ wgpu/examples/hello-windows/main.rs
@@ -15,7 +15,7 @@ struct ViewportDesc {
 
 struct Viewport {
     desc: ViewportDesc,
-    config: wgpu::SurfaceConfiguration,
+    config: wgpu::SurfaceConfiguration<'static>,
 }
 
 impl ViewportDesc {
@@ -39,6 +39,7 @@ impl ViewportDesc {
             height: size.height,
             present_mode: wgpu::PresentMode::Fifo,
             alpha_mode: caps.alpha_modes[0],
+            view_formats: &[],
         };
 
         self.surface.configure(device, &config);